create a topic publisher to constant monitor the signal quality between robot1 and others. the topic output can be further used to creat a wifi receiver topics.
Some of the python scripts are changed slightly.

+++ kheperas_description/python/move.py
@@ -58,7 +58,7 @@ def move():
     
     command = input('~')
     direction = int(command.split(',')[0])
-    distance = int(command.split(',')[1])
+    distance = float(command.split(',')[1])
 
     t = distance / linear_speed * Hz
     t_total = t + Hz
